Rename package, fix tests
